Add files via upload
Updates to post-processing code. When I changed the fertilization input to constant rather than scaled by inflow, it seems like there is not enough P being introduced to the system. Need to correct this issue or revert back to volume-scaled P inputs. Also, 2nd group of chl a is not growing.